refactor: enabled sc versioning for token create
